# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BloomBook API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      parameters:
        - in: query
          name: favorite
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    patch:
      operationId: updateMemory
      tags: [memories]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
  /memories/{id}/favorite:
    patch:
      operationId: toggleMemoryFavorite
      tags: [memories]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"

  # ── CAFES ──────────────────────────────────────────────────────────────
...
  /movies:
    get:
      operationId: listMovies
      tags: [movies]
      responses:
        "200":
          description: Movies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      operationId: createMovie
      tags: [movies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  /movies/{id}:
    get:
      operationId: getMovie
      tags: [movies]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
    patch:
      operationId: updateMovie
      tags: [movies]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
    delete:
      operationId: deleteMovie
      tags: [movies]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted

  # ── WISHLIST ───────────────────────────────────────────────────────────
  /wishlist:
    get:
      operationId: listWishlist
      tags: [wishlist]
      responses:
        "200":
          description: Wishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wish"
    post:
      operationId: createWish
      tags: [wishlist]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wish"
  /wishlist/{id}:
    patch:
      operationId: updateWish
      tags: [wishlist]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wish"
    delete:
      operationId: deleteWish
      tags: [wishlist]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
  /wishlist/{id}/toggle:
    patch:
      operationId: toggleWishDone
      tags: [wishlist]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wish"

  # ── CAPSULES ───────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    patch:
      operationId: updateReview
      tags: [reviews]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Memory:
      type: object
      required: [id, title, isFavorite, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        isFavorite:
          type: ["boolean", "null"]
        createdAt:
          type: string
    MemoryInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        date:
          type: string
        isFavorite:
          type: ["boolean", "null"]
    MemoryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        date:
          type: string
        isFavorite:
          type: ["boolean", "null"]

    # ── Cafe ──────────────────────────────────────────────────────────────
...
    Movie:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        whatItReminded:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    MovieInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        genre:
          type: string
        rating:
          type: integer
        whatItReminded:
          type: string
        posterUrl:
          type: string
    MovieUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        genre:
          type: string
        rating:
          type: integer
        whatItReminded:
          type: string
        posterUrl:
          type: string

    # ── Wish ──────────────────────────────────────────────────────────────
    Wish:
      type: object
      required: [id, title, isDone, createdAt]
      properties:
        id:
          type: string
        emoji:
          type: ["string", "null"]
        title:
          type: string
        notes:
          type: ["string", "null"]
        isDone:
          type: ["boolean", "null"]
        createdAt:
          type: string
    WishInput:
      type: object
      required: [title]
      properties:
        emoji:
          type: string
        title:
          type: string
          minLength: 1
        notes:
          type: string
        isDone:
          type: ["boolean", "null"]
    WishUpdate:
      type: object
      properties:
        emoji:
          type: string
        title:
          type: string
          minLength: 1
        notes:
          type: string
        isDone:
          type: ["boolean", "null"]

    # ── Capsule ───────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, subject, review, createdAt]
      properties:
        id:
          type: string
        subject:
          type: string
        category:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        vibeCheck:
          type: ["string", "null"]
        review:
          type: string
        photoUrl:
          type: ["string", "null"]
        wouldRecommend:
          type: ["boolean", "null"]
        date:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [subject, review]
      properties:
        subject:
          type: string
          minLength: 1
        category:
          type: string
        rating:
          type: integer
        vibeCheck:
          type: string
        review:
          type: string
          minLength: 1
        photoUrl:
          type: string
        wouldRecommend:
          type: ["boolean", "null"]
        date:
          type: string
    ReviewUpdate:
      type: object
      properties:
        subject:
          type: string
          minLength: 1
        category:
          type: string
        rating:
          type: integer
        vibeCheck:
          type: string
        review:
          type: string
          minLength: 1
        photoUrl:
          type: string
        wouldRecommend:
          type: ["boolean", "null"]
        date:
          type: string