# lib/api-zod/src/generated/types/memoryUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  mediaUrl?: string;
  mediaType?: string;
  date?: string;
  /** @nullable */
  isFavorite?: boolean | null;
}