# lib/api-zod/src/generated/types/movieInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MovieInput {
  /** @minLength 1 */
  title: string;
  type?: string;
  genre?: string;
  rating?: number;
  whatItReminded?: string;
  posterUrl?: string;
}