# lib/api-zod/src/generated/types/reviewUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BloomBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewUpdate {
  /** @minLength 1 */
  subject?: string;
  category?: string;
  rating?: number;
  vibeCheck?: string;
  /** @minLength 1 */
  review?: string;
  photoUrl?: string;
  /** @nullable */
  wouldRecommend?: boolean | null;
  date?: string;
}